Linear_Regression
In this, you will see the implementation of the Linear_Regression, in Python language, with the help of Jupiter notebook ide. I have already uploaded the excel file called supermarket_sales.xlsx. In this, we have to predict the Outlet sales. You will find all the related information, related to the Linear_Regression algorithm.